+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format of this change log follows the advice given at [Keep a CHANGELOG](htt
 
 - Add support for Rule in the following rules:
   - keywords-in-logical-order
+  - no-files-without-scenarios
 
 ###  Changed